environment: introduce definition builder
